Renames Change to ChangeValue and implements abstract basetype Change.
